ffi: many fixes caught by pyo3-ffi-check
