+++ pyo3-ffi/src/compile.rs
@@ -5,4 +5,6 @@ pub const Py_file_input: c_int = 257;
 pub const Py_eval_input: c_int = 258;
 #[cfg(Py_3_8)]
 pub const Py_func_type_input: c_int = 345;
-// skipped Py_fstring_input
+
+#[cfg(Py_3_9)]
+pub const Py_fstring_input: c_int = 800;
