+++ pyo3-ffi/src/cpython/object.rs
@@ -1,5 +1,7 @@
-use crate::PyObject;
-use std::os::raw::c_int;
+use crate::object;
+use crate::{PyObject, Py_ssize_t};
+use std::mem;
+use std::os::raw::{c_char, c_int, c_uint, c_ulong, c_void};
 
 // skipped _Py_NewReference
 // skipped _Py_ForgetReference
@@ -118,209 +120,208 @@ pub type vectorcallfunc = unsafe extern "C" fn(
     kwnames: *mut PyObject,
 ) -> *mut PyObject;
 
-mod typeobject {
-    use crate::object;
-    use crate::{PyObject, Py_ssize_t};
-    use std::mem;
-    use std::os::raw::{c_char, c_int, c_uint, c_ulong, c_void};
+#[repr(C)]
+#[derive(Copy, Clone)]
+pub struct PyNumberMethods {
+    pub nb_add: Option<object::binaryfunc>,
+    pub nb_subtract: Option<object::binaryfunc>,
+    pub nb_multiply: Option<object::binaryfunc>,
+    pub nb_remainder: Option<object::binaryfunc>,
+    pub nb_divmod: Option<object::binaryfunc>,
+    pub nb_power: Option<object::ternaryfunc>,
+    pub nb_negative: Option<object::unaryfunc>,
+    pub nb_positive: Option<object::unaryfunc>,
+    pub nb_absolute: Option<object::unaryfunc>,
+    pub nb_bool: Option<object::inquiry>,
+    pub nb_invert: Option<object::unaryfunc>,
+    pub nb_lshift: Option<object::binaryfunc>,
+    pub nb_rshift: Option<object::binaryfunc>,
+    pub nb_and: Option<object::binaryfunc>,
+    pub nb_xor: Option<object::binaryfunc>,
+    pub nb_or: Option<object::binaryfunc>,
+    pub nb_int: Option<object::unaryfunc>,
+    pub nb_reserved: *mut c_void,
+    pub nb_float: Option<object::unaryfunc>,
+    pub nb_inplace_add: Option<object::binaryfunc>,
+    pub nb_inplace_subtract: Option<object::binaryfunc>,
+    pub nb_inplace_multiply: Option<object::binaryfunc>,
+    pub nb_inplace_remainder: Option<object::binaryfunc>,
+    pub nb_inplace_power: Option<object::ternaryfunc>,
+    pub nb_inplace_lshift: Option<object::binaryfunc>,
+    pub nb_inplace_rshift: Option<object::binaryfunc>,
+    pub nb_inplace_and: Option<object::binaryfunc>,
+    pub nb_inplace_xor: Option<object::binaryfunc>,
+    pub nb_inplace_or: Option<object::binaryfunc>,
+    pub nb_floor_divide: Option<object::binaryfunc>,
+    pub nb_true_divide: Option<object::binaryfunc>,
+    pub nb_inplace_floor_divide: Option<object::binaryfunc>,
+    pub nb_inplace_true_divide: Option<object::binaryfunc>,
+    pub nb_index: Option<object::unaryfunc>,
+    pub nb_matrix_multiply: Option<object::binaryfunc>,
+    pub nb_inplace_matrix_multiply: Option<object::binaryfunc>,
+}
 
-    #[repr(C)]
-    #[derive(Copy, Clone)]
-    pub struct PyNumberMethods {
-        pub nb_add: Option<object::binaryfunc>,
-        pub nb_subtract: Option<object::binaryfunc>,
-        pub nb_multiply: Option<object::binaryfunc>,
-        pub nb_remainder: Option<object::binaryfunc>,
-        pub nb_divmod: Option<object::binaryfunc>,
-        pub nb_power: Option<object::ternaryfunc>,
-        pub nb_negative: Option<object::unaryfunc>,
-        pub nb_positive: Option<object::unaryfunc>,
-        pub nb_absolute: Option<object::unaryfunc>,
-        pub nb_bool: Option<object::inquiry>,
-        pub nb_invert: Option<object::unaryfunc>,
-        pub nb_lshift: Option<object::binaryfunc>,
-        pub nb_rshift: Option<object::binaryfunc>,
-        pub nb_and: Option<object::binaryfunc>,
-        pub nb_xor: Option<object::binaryfunc>,
-        pub nb_or: Option<object::binaryfunc>,
-        pub nb_int: Option<object::unaryfunc>,
-        pub nb_reserved: *mut c_void,
-        pub nb_float: Option<object::unaryfunc>,
-        pub nb_inplace_add: Option<object::binaryfunc>,
-        pub nb_inplace_subtract: Option<object::binaryfunc>,
-        pub nb_inplace_multiply: Option<object::binaryfunc>,
-        pub nb_inplace_remainder: Option<object::binaryfunc>,
-        pub nb_inplace_power: Option<object::ternaryfunc>,
-        pub nb_inplace_lshift: Option<object::binaryfunc>,
-        pub nb_inplace_rshift: Option<object::binaryfunc>,
-        pub nb_inplace_and: Option<object::binaryfunc>,
-        pub nb_inplace_xor: Option<object::binaryfunc>,
-        pub nb_inplace_or: Option<object::binaryfunc>,
-        pub nb_floor_divide: Option<object::binaryfunc>,
-        pub nb_true_divide: Option<object::binaryfunc>,
-        pub nb_inplace_floor_divide: Option<object::binaryfunc>,
-        pub nb_inplace_true_divide: Option<object::binaryfunc>,
-        pub nb_index: Option<object::unaryfunc>,
-        pub nb_matrix_multiply: Option<object::binaryfunc>,
-        pub nb_inplace_matrix_multiply: Option<object::binaryfunc>,
-    }
+#[repr(C)]
+#[derive(Clone)]
+pub struct PySequenceMethods {
+    pub sq_length: Option<object::lenfunc>,
+    pub sq_concat: Option<object::binaryfunc>,
+    pub sq_repeat: Option<object::ssizeargfunc>,
+    pub sq_item: Option<object::ssizeargfunc>,
+    pub was_sq_slice: *mut c_void,
+    pub sq_ass_item: Option<object::ssizeobjargproc>,
+    pub was_sq_ass_slice: *mut c_void,
+    pub sq_contains: Option<object::objobjproc>,
+    pub sq_inplace_concat: Option<object::binaryfunc>,
+    pub sq_inplace_repeat: Option<object::ssizeargfunc>,
+}
 
-    #[repr(C)]
-    #[derive(Clone)]
-    pub struct PySequenceMethods {
-        pub sq_length: Option<object::lenfunc>,
-        pub sq_concat: Option<object::binaryfunc>,
-        pub sq_repeat: Option<object::ssizeargfunc>,
-        pub sq_item: Option<object::ssizeargfunc>,
-        pub was_sq_slice: *mut c_void,
-        pub sq_ass_item: Option<object::ssizeobjargproc>,
-        pub was_sq_ass_slice: *mut c_void,
-        pub sq_contains: Option<object::objobjproc>,
-        pub sq_inplace_concat: Option<object::binaryfunc>,
-        pub sq_inplace_repeat: Option<object::ssizeargfunc>,
-    }
+#[repr(C)]
+#[derive(Clone, Default)]
+pub struct PyMappingMethods {
+    pub mp_length: Option<object::lenfunc>,
+    pub mp_subscript: Option<object::binaryfunc>,
+    pub mp_ass_subscript: Option<object::objobjargproc>,
+}
 
-    #[repr(C)]
-    #[derive(Clone, Default)]
-    pub struct PyMappingMethods {
-        pub mp_length: Option<object::lenfunc>,
-        pub mp_subscript: Option<object::binaryfunc>,
-        pub mp_ass_subscript: Option<object::objobjargproc>,
-    }
+#[cfg(Py_3_10)]
+pub type sendfunc = unsafe extern "C" fn(
+    iter: *mut PyObject,
+    value: *mut PyObject,
+    result: *mut *mut PyObject,
+) -> object::PySendResult;
 
-    // skipped PySendResult
+#[repr(C)]
+#[derive(Clone, Default)]
+pub struct PyAsyncMethods {
+    pub am_await: Option<object::unaryfunc>,
+    pub am_aiter: Option<object::unaryfunc>,
+    pub am_anext: Option<object::unaryfunc>,
+    #[cfg(Py_3_10)]
+    pub am_send: Option<sendfunc>,
+}
 
-    #[repr(C)]
-    #[derive(Clone, Default)]
-    pub struct PyAsyncMethods {
-        pub am_await: Option<object::unaryfunc>,
-        pub am_aiter: Option<object::unaryfunc>,
-        pub am_anext: Option<object::unaryfunc>,
-    }
+#[repr(C)]
+#[derive(Clone, Default)]
+pub struct PyBufferProcs {
+    pub bf_getbuffer: Option<crate::getbufferproc>,
+    pub bf_releasebuffer: Option<crate::releasebufferproc>,
+}
 
-    #[repr(C)]
-    #[derive(Clone, Default)]
-    pub struct PyBufferProcs {
-        pub bf_getbuffer: Option<crate::getbufferproc>,
-        pub bf_releasebuffer: Option<crate::releasebufferproc>,
-    }
+pub type printfunc =
+    unsafe extern "C" fn(arg1: *mut PyObject, arg2: *mut ::libc::FILE, arg3: c_int) -> c_int;
 
-    pub type printfunc =
-        unsafe extern "C" fn(arg1: *mut PyObject, arg2: *mut ::libc::FILE, arg3: c_int) -> c_int;
+#[repr(C)]
+#[derive(Debug, Copy, Clone)]
+pub struct PyTypeObject {
+    #[cfg(PyPy)]
+    pub ob_refcnt: Py_ssize_t,
+    #[cfg(PyPy)]
+    pub ob_pypy_link: Py_ssize_t,
+    #[cfg(PyPy)]
+    pub ob_type: *mut PyTypeObject,
+    #[cfg(PyPy)]
+    pub ob_size: Py_ssize_t,
+    #[cfg(not(PyPy))]
+    pub ob_base: object::PyVarObject,
+    pub tp_name: *const c_char,
+    pub tp_basicsize: Py_ssize_t,
+    pub tp_itemsize: Py_ssize_t,
+    pub tp_dealloc: Option<object::destructor>,
+    #[cfg(not(Py_3_8))]
+    pub tp_print: Option<printfunc>,
+    #[cfg(Py_3_8)]
+    pub tp_vectorcall_offset: Py_ssize_t,
+    pub tp_getattr: Option<object::getattrfunc>,
+    pub tp_setattr: Option<object::setattrfunc>,
+    pub tp_as_async: *mut PyAsyncMethods,
+    pub tp_repr: Option<object::reprfunc>,
+    pub tp_as_number: *mut PyNumberMethods,
+    pub tp_as_sequence: *mut PySequenceMethods,
+    pub tp_as_mapping: *mut PyMappingMethods,
+    pub tp_hash: Option<object::hashfunc>,
+    pub tp_call: Option<object::ternaryfunc>,
+    pub tp_str: Option<object::reprfunc>,
+    pub tp_getattro: Option<object::getattrofunc>,
+    pub tp_setattro: Option<object::setattrofunc>,
+    pub tp_as_buffer: *mut PyBufferProcs,
+    pub tp_flags: c_ulong,
+    pub tp_doc: *const c_char,
+    pub tp_traverse: Option<object::traverseproc>,
+    pub tp_clear: Option<object::inquiry>,
+    pub tp_richcompare: Option<object::richcmpfunc>,
+    pub tp_weaklistoffset: Py_ssize_t,
+    pub tp_iter: Option<object::getiterfunc>,
+    pub tp_iternext: Option<object::iternextfunc>,
+    pub tp_methods: *mut crate::methodobject::PyMethodDef,
+    pub tp_members: *mut crate::structmember::PyMemberDef,
+    pub tp_getset: *mut crate::descrobject::PyGetSetDef,
+    pub tp_base: *mut PyTypeObject,
+    pub tp_dict: *mut object::PyObject,
+    pub tp_descr_get: Option<object::descrgetfunc>,
+    pub tp_descr_set: Option<object::descrsetfunc>,
+    pub tp_dictoffset: Py_ssize_t,
+    pub tp_init: Option<object::initproc>,
+    pub tp_alloc: Option<object::allocfunc>,
+    pub tp_new: Option<object::newfunc>,
+    pub tp_free: Option<object::freefunc>,
+    pub tp_is_gc: Option<object::inquiry>,
+    pub tp_bases: *mut object::PyObject,
+    pub tp_mro: *mut object::PyObject,
+    pub tp_cache: *mut object::PyObject,
+    pub tp_subclasses: *mut object::PyObject,
+    pub tp_weaklist: *mut object::PyObject,
+    pub tp_del: Option<object::destructor>,
+    pub tp_version_tag: c_uint,
+    pub tp_finalize: Option<object::destructor>,
+    #[cfg(Py_3_8)]
+    pub tp_vectorcall: Option<super::vectorcallfunc>,
+    #[cfg(PyPy)]
+    pub tp_pypy_flags: std::os::raw::c_long,
+    #[cfg(py_sys_config = "COUNT_ALLOCS")]
+    pub tp_allocs: Py_ssize_t,
+    #[cfg(py_sys_config = "COUNT_ALLOCS")]
+    pub tp_frees: Py_ssize_t,
+    #[cfg(py_sys_config = "COUNT_ALLOCS")]
+    pub tp_maxalloc: Py_ssize_t,
+    #[cfg(py_sys_config = "COUNT_ALLOCS")]
+    pub tp_prev: *mut PyTypeObject,
+    #[cfg(py_sys_config = "COUNT_ALLOCS")]
+    pub tp_next: *mut PyTypeObject,
+}
 
-    #[repr(C)]
-    #[derive(Debug, Copy, Clone)]
-    pub struct PyTypeObject {
-        #[cfg(PyPy)]
-        pub ob_refcnt: Py_ssize_t,
-        #[cfg(PyPy)]
-        pub ob_pypy_link: Py_ssize_t,
-        #[cfg(PyPy)]
-        pub ob_type: *mut PyTypeObject,
-        #[cfg(PyPy)]
-        pub ob_size: Py_ssize_t,
-        #[cfg(not(PyPy))]
-        pub ob_base: object::PyVarObject,
-        pub tp_name: *const c_char,
-        pub tp_basicsize: Py_ssize_t,
-        pub tp_itemsize: Py_ssize_t,
-        pub tp_dealloc: Option<object::destructor>,
-        #[cfg(not(Py_3_8))]
-        pub tp_print: Option<printfunc>,
-        #[cfg(Py_3_8)]
-        pub tp_vectorcall_offset: Py_ssize_t,
-        pub tp_getattr: Option<object::getattrfunc>,
-        pub tp_setattr: Option<object::setattrfunc>,
-        pub tp_as_async: *mut PyAsyncMethods,
-        pub tp_repr: Option<object::reprfunc>,
-        pub tp_as_number: *mut PyNumberMethods,
-        pub tp_as_sequence: *mut PySequenceMethods,
-        pub tp_as_mapping: *mut PyMappingMethods,
-        pub tp_hash: Option<object::hashfunc>,
-        pub tp_call: Option<object::ternaryfunc>,
-        pub tp_str: Option<object::reprfunc>,
-        pub tp_getattro: Option<object::getattrofunc>,
-        pub tp_setattro: Option<object::setattrofunc>,
-        pub tp_as_buffer: *mut PyBufferProcs,
-        pub tp_flags: c_ulong,
-        pub tp_doc: *const c_char,
-        pub tp_traverse: Option<object::traverseproc>,
-        pub tp_clear: Option<object::inquiry>,
-        pub tp_richcompare: Option<object::richcmpfunc>,
-        pub tp_weaklistoffset: Py_ssize_t,
-        pub tp_iter: Option<object::getiterfunc>,
-        pub tp_iternext: Option<object::iternextfunc>,
-        pub tp_methods: *mut crate::methodobject::PyMethodDef,
-        pub tp_members: *mut crate::structmember::PyMemberDef,
-        pub tp_getset: *mut crate::descrobject::PyGetSetDef,
-        pub tp_base: *mut PyTypeObject,
-        pub tp_dict: *mut object::PyObject,
-        pub tp_descr_get: Option<object::descrgetfunc>,
-        pub tp_descr_set: Option<object::descrsetfunc>,
-        pub tp_dictoffset: Py_ssize_t,
-        pub tp_init: Option<object::initproc>,
-        pub tp_alloc: Option<object::allocfunc>,
-        pub tp_new: Option<object::newfunc>,
-        pub tp_free: Option<object::freefunc>,
-        pub tp_is_gc: Option<object::inquiry>,
-        pub tp_bases: *mut object::PyObject,
-        pub tp_mro: *mut object::PyObject,
-        pub tp_cache: *mut object::PyObject,
-        pub tp_subclasses: *mut object::PyObject,
-        pub tp_weaklist: *mut object::PyObject,
-        pub tp_del: Option<object::destructor>,
-        pub tp_version_tag: c_uint,
-        pub tp_finalize: Option<object::destructor>,
-        #[cfg(Py_3_8)]
-        pub tp_vectorcall: Option<super::vectorcallfunc>,
-        #[cfg(PyPy)]
-        pub tp_pypy_flags: std::os::raw::c_long,
-        #[cfg(py_sys_config = "COUNT_ALLOCS")]
-        pub tp_allocs: Py_ssize_t,
-        #[cfg(py_sys_config = "COUNT_ALLOCS")]
-        pub tp_frees: Py_ssize_t,
-        #[cfg(py_sys_config = "COUNT_ALLOCS")]
-        pub tp_maxalloc: Py_ssize_t,
-        #[cfg(py_sys_config = "COUNT_ALLOCS")]
-        pub tp_prev: *mut PyTypeObject,
-        #[cfg(py_sys_config = "COUNT_ALLOCS")]
-        pub tp_next: *mut PyTypeObject,
-    }
+#[repr(C)]
+#[derive(Clone)]
+pub struct PyHeapTypeObject {
+    pub ht_type: PyTypeObject,
+    pub as_async: PyAsyncMethods,
+    pub as_number: PyNumberMethods,
+    pub as_mapping: PyMappingMethods,
+    pub as_sequence: PySequenceMethods,
+    pub as_buffer: PyBufferProcs,
+    pub ht_name: *mut object::PyObject,
+    pub ht_slots: *mut object::PyObject,
+    pub ht_qualname: *mut object::PyObject,
+    pub ht_cached_keys: *mut c_void,
+    #[cfg(Py_3_9)]
+    pub ht_module: *mut object::PyObject,
+}
 
-    #[repr(C)]
-    #[derive(Clone)]
-    pub struct PyHeapTypeObject {
-        pub ht_type: PyTypeObject,
-        pub as_async: PyAsyncMethods,
-        pub as_number: PyNumberMethods,
-        pub as_mapping: PyMappingMethods,
-        pub as_sequence: PySequenceMethods,
-        pub as_buffer: PyBufferProcs,
-        pub ht_name: *mut object::PyObject,
-        pub ht_slots: *mut object::PyObject,
-        pub ht_qualname: *mut object::PyObject,
-        pub ht_cached_keys: *mut c_void,
-    }
-
-    impl Default for PyHeapTypeObject {
-        #[inline]
-        fn default() -> Self {
-            unsafe { mem::zeroed() }
-        }
-    }
-
+impl Default for PyHeapTypeObject {
     #[inline]
-    pub unsafe fn PyHeapType_GET_MEMBERS(
-        etype: *mut PyHeapTypeObject,
-    ) -> *mut crate::structmember::PyMemberDef {
-        let py_type = object::Py_TYPE(etype as *mut object::PyObject);
-        let ptr = etype.offset((*py_type).tp_basicsize);
-        ptr as *mut crate::structmember::PyMemberDef
+    fn default() -> Self {
+        unsafe { mem::zeroed() }
     }
 }
 
-// The exported types depend on whether Py_LIMITED_API is set
-pub use self::typeobject::*;
+#[inline]
+pub unsafe fn PyHeapType_GET_MEMBERS(
+    etype: *mut PyHeapTypeObject,
+) -> *mut crate::structmember::PyMemberDef {
+    let py_type = object::Py_TYPE(etype as *mut object::PyObject);
+    let ptr = etype.offset((*py_type).tp_basicsize);
+    ptr as *mut crate::structmember::PyMemberDef
+}
 
 // skipped _PyType_Name
 // skipped _PyType_Lookup
