+++ pyo3-ffi/src/cpython/pystate.rs
@@ -26,7 +26,16 @@ pub const PyTrace_OPCODE: c_int = 7;
 
 // skipped PyTraceInfo
 // skipped CFrame
-// skipped _PyErr_StackItem
+
+#[repr(C)]
+#[derive(Clone, Copy)]
+pub struct _PyErr_StackItem {
+    pub exc_type: *mut PyObject,
+    pub exc_value: *mut PyObject,
+    pub exc_traceback: *mut PyObject,
+    pub previous_item: *mut _PyErr_StackItem,
+}
+
 // skipped _PyStackChunk
 // skipped _ts (aka PyThreadState)
